GEN-403 Get audio working on Raspberry Pi
Audio is playing mostly OK on the Pi 2 B now.

At times of high CPU usage (such as when the logo first paints) there are still the occasional ALSO buffer
underruns:

```
ALSA lib pcm.c:8306:(snd_pcm_recover) underrun occurred
```

We might be able to avoid this if we can figure out how to shrink the size of the SDL audio buffers a bit.